Refactor: Moved peer review helper methods into a helper class
